chore: move workshop & student-list (from modules.tsx to modules)
