+++ src/modules/dashboard/nutritionist/workshop/CourseBigCard.tsx
