+++ src/modules/dashboard/nutritionist/workshop/CourseDeleteModal.tsx
