+++ src/pages/dashboard/nutritionist/workshop/courses/index.tsx
@@ -1,6 +1,6 @@
 import DashboardOuterContainer from "@/common/components/DashboardOuterContainer";
-import CourseBigCard from "@/modules.tsx/dashboard/nutritionist/CourseBigCard";
 import NutritionistDashboardLayout from "@/modules.tsx/dashboard/nutritionist/DashboardLayout";
+import CourseBigCard from "@/modules/dashboard/nutritionist/workshop/CourseBigCard";
 import Link from "next/link";
 
 const NutritionistCoursePage = () => {
